+++ src/main/java/com/greenvn/starlightelectronicsstore/repository/ProductAttributeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.greenvn.starlightelectronicsstore.entities.ProductAttribute;
 
 @Repository
-public interface ProductAttributeRepository extends JpaRepository<ProductAttribute,Integer>{
+public interface ProductAttributeRepository extends JpaRepository<ProductAttribute,Long>{
 
 }
